+++ src/main/java/com/zerobase/moy/data/model/sign/ReissueResultDto.java
@@ -1,21 +1,18 @@
-package com.zerobase.moy.data.model;
+package com.zerobase.moy.data.model.sign;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 import lombok.ToString;
 
 @NoArgsConstructor
 @AllArgsConstructor
 @ToString
 @Getter
-@Setter
-public class SignUpResultDto {
+@Builder
+public class ReissueResultDto {
 
-  private boolean success;
+  TokenDto token;
 
-  private int code;
-
-  private String msg;
 }
